Fixed my rounding stupids

+++ src/main/java/redd90/exprimo/essentia/IEssentiaContainer.java
@@ -6,7 +6,7 @@ public interface IEssentiaContainer {
 	
 	StackSet getStackSet();
 	
-	int getInnerPressure(Essentia essentia);
+	double getInnerPressure(Essentia essentia);
 
 
 }
